strip down a bit- blob http server and basic window behaviors
