+++ app/lib/windows.js
@@ -2,7 +2,6 @@ var BrowserWindow = require('browser-window')
 var path = require('path')
 var shell = require('shell')
 var setupRpc = require('./muxrpc-ipc')
-var blobslib = require('./blobs')
 
 var windows = []
 
@@ -20,7 +19,7 @@ var secureWebPreferences = {
 }
 
 var open =
-module.exports.open = function (url, sbot, blobs, opts, params) {
+module.exports.open = function (sbot, url, opts, params) {
   var win = new BrowserWindow(opts)
   win.loadUrl(url)
   setupRpc(win, sbot, params)
@@ -34,42 +33,8 @@ module.exports.open = function (url, sbot, blobs, opts, params) {
   
   win.webContents.on('new-window', function (e, url) {
     e.preventDefault() // hell naw
-    if (url.indexOf('http://localhost:7777/') === 0) {
-      // open the file
-      blobs.checkout(url, function (err, filepath) {
-        if (err) {
-          if (err.badUrl)
-            alert('Error: Not a valid file reference')
-          else if (err.notFound) {
-            // register want
-            var hash = blobslib.url_parse(url).hash
-            sbot.blobs.want(hash, {nowait: true}, function(){})
-            // open search interface
-            open(
-              'file://' + path.join(__dirname, '../../node_modules/ssbplug-phoenix/blob-search.html'),
-              sbot,
-              blobs,
-              { width: 600, height: 650, 'web-preferences': secureWebPreferences },
-              { url: url, hash: hash }
-            )
-          } else
-            console.log(err) // :TODO: something nicer
-        } else {
-          shell.openItem(filepath) // open in desktop's default program
-        }
-      })
-    } else if (url.indexOf('file://') === 0) {
-      var mainWindow = open(
-        url,
-        sbot,
-        blobs,
-        { width: 1030, height: 720 }
-      )
-      require('./menu')(mainWindow)
-    } else {
-      // open in the browser
-      shell.openExternal(url)
-    }
+    // open in the browser
+    shell.openExternal(url)
   })
 
   return win
